Support x86 and arm64 in tests
This adds x86 and arm64 versions of the "text" extension to make GitHub Actions work again. The `macos-latest` GitHub Actions runner is only available as arm64. It also adds `macos-13` to the test matrix so that x86 is tested on macOS as well.

The extensions were downloaded from https://github.com/nalgeon/sqlean